Improves the speed of reflex CLI commands with lazy imports.

Before:
```
~/code/reflex main ❯ hyperfine "poetry run reflex -v"                                                                                                                                                                                             
Benchmark 1: poetry run reflex -v                                                                                                                                                                                                                 
  Time (mean ± σ):     860.1 ms ± 268.8 ms    [User: 574.2 ms, System: 132.2 ms]                                                                                                                                                                  
  Range (min … max):   746.4 ms … 1623.6 ms    10 runs   
```

After:
```
~/code/reflex nikhil/lazy2 ❯ hyperfine "poetry run reflex -v"                                                                                                                                                                                     
Benchmark 1: poetry run reflex -v                                                                                                                                                                                                                 
  Time (mean ± σ):     478.0 ms ±  17.0 ms    [User: 320.9 ms, System: 86.3 ms]                                          
  Range (min … max):   460.2 ms … 516.4 ms    10 runs
```